support an arbitrary tag on the reported metrics
The `STATSD_GROUPING` environment variable adds a `grouping` tag to the
metrics, with a value equal to the value of the environment variable.
This is particularly helpful in a kubernetes deployment where each pod
has a unique name but you want the option to group metrics across all
pods in a deployment. Setting this on the pods in a deployment might
look something like:

```yaml
env:
  - name: STATSD_GROUPING
    value: deployment-foo
```