remove json from the gemspec
Now that json is a default gem I liked the idea of explictly listed json
as a requirement, however including it is causing grief. When puma
starts in some environments and the Gemfile.lock has a version of json
that conflicts with the default gem version shipped with the local ruby
then users get this sort of warning (if they're lucky. Sometimes they
get an error):

    [2437] Puma starting in cluster mode...
    [2437] * Version 4.3.6 (ruby 2.6.6-p146), codename: Mysterious Traveller
    [2437] * Min threads: 32, max threads: 32
    [2437] * Environment: production
    [2437] * Process workers: 2
    [2437] * Phased restart available
    [2437] * Listening on unix:///app/tmp/sockets/puma.sock?backlog=1024
    [2437] Use Ctrl-C to stop
    [2464] + Gemfile in context: /app/Gemfile
    [2466] + Gemfile in context: /app/Gemfile
    WARN: Unresolved or ambigious specs during Gem::Specification.reset:
          json (>= 0)
    WARN: Unresolved or ambigious specs during Gem::Specification.reset:
          json (>= 0)
          Available/installed versions of this gem:
          - 2.3.0
          - 2.1.0
    WARN: Clearing out unresolved specs. Try 'gem cleanup <gem>'
    Please report a bug if this causes problems.
          Available/installed versions of this gem:
          - 2.3.0
          - 2.1.0
    WARN: Clearing out unresolved specs. Try 'gem cleanup <gem>'
    Please report a bug if this causes problems.

We don't strictly need it listed as a dependency. As a default gem, a
plain `require 'json'` will work just fine on all rubies we target.