require puma 5 or better
puma 5 added a new API for fetching the stats as a plain ruby Hash, with
no need to deserialise JSON.

Avoiding JSON is nice, because the json gem is a "default gem" that
loads differently and it's difficult to require without load order
issues.

I plan to release this as v2.x, so folks on older puma versions can
continue using the 1.x series.