Advanced configuration
There are situations when you want to be able to configure plugin using a Ruby
code:

1. Sometimes you already have environment variables with the values you need,
just with different names, and want to avoid duplication.
2. Sometimes you want to use ruby to calculate the configuration from other
environment variables or whatnot.
3. Sometimes you want to enable plugin conditionally, and conditionally
configure it.

All three reasons we experienced in prod and currently have to maintain a fork
with the Ruby configuration. The code we suggest is tried and true in prod over
three months at the moment, and while we are here, offering the tests for
unchanged code as well.