Add new puma.requests metric
Confusingly, the puma metric called "requests_count" is actually a gauge
in statsd terms. It's the sum of all requests that have been served
since this puma was booted, always increasing.

Confusing name aside, a gauge of total requests served is useful.
However, it can always be useful to report the delta as a statsd
counter. That is, the number of requests served since the last time we
reported (by default, 2 seconds ago).

It was tempting to change puma.requests_count to be the statsd counter
and move the gauge to a new name. That would be a breaking change
though, and I wasn't willing to do that.