Add percent_busy metric
This metric is helpful to determine how loaded the puma instances are.

It's a simple calculation derived from existing metrics so it could be easily
calculated in the consumer of the metrics. Unfortunately when using Datadog and
Kubernetes horizontal pod autoscaling it's not possible to derive or calculate
a metric, the metric must be materialized as a regular metric in Datadog.